Change Query module style to a more ES6-ish pattern
The class pattern used was a bit strange since we have a collection
of library functions that don't need to share an "instance" of a class.
You can use "import * as ExampleQuery from './query/exampleQuery'".
This works more like a traditional language where you import
library functions under a namespace.